The thread pool size should be nbOfThreads * nb of KafkaReader implementations.
Note that all instances will be part of the same consumer group today.
This is perfectly fine as each thread will only subscribe to it's own topics.
If you want something more fine grain, you probably need to move the creation of the KafkaConsumer within the KafkaReader class instead of autowiring it.